Project3 : Django RDB V1.0

+++ Project2/crud/articles/views.py
@@ -1,7 +1,7 @@
 # articles/views.py
 
-from .models import Article
-from .forms import ArticleForm
+from .models import Article, Comment
+from .forms import ArticleForm, CommentForm
 from django.shortcuts import render, redirect
 from django.views.decorators.http import require_http_methods, require_POST, require_safe
 from django.contrib.auth.decorators import login_required
@@ -17,6 +17,8 @@ def create(request):
     if request.method == 'POST':
         form = ArticleForm(request.POST)
         if form.is_valid():
+            article = form.save(commit=False)
+            article.user = request.user
             article = form.save()
             return redirect('articles:detail', article.pk)
     else:
@@ -29,32 +31,66 @@ def create(request):
 @require_safe
 def detail(request, pk):
     article = Article.objects.get(pk=pk)
+    comment_form = CommentForm()
     context = {
         'article' : article,
+        'comment_form' : comment_form,
     }
     return render(request, 'articles/detail.html', context)
 
 @login_required
 @require_POST
 def delete(request, pk):
+    article = Article.objects.get(pk=pk)
     if request.user.is_authenticated:
-        article = Article.objects.get(pk=pk)
-        article.delete()
-    return redirect('articles:index')
+        if request.user == article.user:
+            article = Article.objects.get(pk=pk)
+            article.delete()
+            return redirect('articles:index')
+    return redirect('articles:detail', article.pk)
 
 @login_required
 @require_http_methods(['GET', 'POST'])
 def update(request, pk):
     article = Article.objects.get(pk=pk)
-    if request.method == 'POST':
-        form = ArticleForm(request.POST, instance=article)
-        if form.is_valid():
-            form.save()
-            return redirect('articles:detail', article.pk)
+    if request.user == article.user:
+        if request.method == 'POST':
+            form = ArticleForm(request.POST, request.FILES, instance=article)
+            if form.is_valid():
+                form.save()
+                return redirect('articles:detail', article.pk)
+        else:
+            form = ArticleForm(instance=article)
     else:
-        form = ArticleForm(instance=article)
+        return redirect('articles:index')
     context = {
-        'form' : form,
-        'article' : article,
+        'article': article,
+        'form': form,
     }
     return render(request, 'articles/update.html', context)
+
+
+def comments_create(request, pk):
+    article = Article.objects.get(pk = pk)
+    comment_form = CommentForm(request.POST)
+    if comment_form.is_valid():
+        comment = comment_form.save(commit=False)
+        comment.article = article
+        comment.save()
+    return redirect('articles:detail', article.pk)
+
+def detail(request, pk):
+    article = Article.objects.get(pk=pk)
+    comment_form = CommentForm()
+    comments = article.comment_set.all()
+    context = {
+        'article' : article,
+        'comment_form' : comment_form,
+        'comments' : comments,
+    }
+    return render(request, 'articles/detail.html', context)
+
+def comments_delete(request, article_pk, comment_pk):
+    comment = Comment.objects.get(pk=comment_pk)
+    comment.delete()
+    return redirect('articles:detail', article_pk)
